Extract stream-based HoC
Maps the provided stream of props to an element along with
the emitter.
