+++ packages/brookjs-cli/src/deltas/terminalDelta/index.js
@@ -1,32 +1,10 @@
 import { Kefir, ofType } from 'brookjs';
-import { h, Component, render } from 'ink';
 import { RUN } from '../../actions';
 import log from './log';
 import CommandNotFound from './CommandNotFound';
 import devCommand from './devCommand';
 import newCommandPrompt from './newCommandPrompt';
-
-class Root extends Component {
-    componentDidMount() {
-        const { props$ } = this.props;
-
-        this.sub = props$.observe(val => {
-            this.setState(val);
-        });
-    }
-
-    componentWillUnmount() {
-        this.sub && this.sub.unsubscribe();
-    }
-
-    render({ emitter, View }) {
-        return <View emitter={emitter} {...this.state} />;
-    }
-}
-
-const mountTerminal = (View, props$) => Kefir.stream(emitter =>
-    render(<Root props$={props$} View={View} emitter={emitter}/>)
-);
+import mountTerminal from './mountTerminal';
 
 export default (services) => (actions$, state$) =>
     Kefir.merge([
